API change in r3114 broke the indian module

+++ dhis-web-dataentry-linelisting/src/main/java/org/hisp/dhis/den/action/SectionFormAction.java
@@ -259,10 +259,10 @@ public class SectionFormAction
 
         calculatedValueMap = new HashMap<CalculatedDataElement, Integer>();
 
-        Map<DataElement, Integer> factorMap;
+        Map<DataElement, Double> factorMap;
 
         DataValue dataValue;
-        int factor;
+        Double factor;
         int value = 0;
 
         for ( DataElement dataElement : dataElements )
