+++ dhis-web-dataentry-linelisting/src/main/java/org/hisp/dhis/den/action/multidimensional/SelectAction.java
@@ -165,9 +165,9 @@ public class SelectAction
         return calculatedDataElementIds;
     }
 
-    private Map<CalculatedDataElement, Map<DataElement, Integer>> calculatedDataElementMap;
+    private Map<CalculatedDataElement, Map<DataElement, Double>> calculatedDataElementMap;
 
-    public Map<CalculatedDataElement, Map<DataElement, Integer>> getCalculatedDataElementMap()
+    public Map<CalculatedDataElement, Map<DataElement, Double>> getCalculatedDataElementMap()
     {
         return calculatedDataElementMap;
     }
@@ -320,7 +320,7 @@ public class SelectAction
         // ---------------------------------------------------------------------
 
         calculatedDataElementIds = new HashSet<Integer>();
-        calculatedDataElementMap = new HashMap<CalculatedDataElement, Map<DataElement, Integer>>();
+        calculatedDataElementMap = new HashMap<CalculatedDataElement, Map<DataElement, Double>>();
         CalculatedDataElement cde;
 
         for ( DataElement dataElement : selectedDataSet.getDataElements() )
